+++ bot/utilities/cleverbot.js
@@ -15,7 +15,12 @@ module.exports = function(bot, db, data) {
   Cleverbot.prepare(function(){
     cleverbot.write(data.params.join(" "), function (response) {
       if (response) {
+        if (response.sessionid.indexOf('404') >= 0 ) {
+          bot.log('error', 'BOT', '404 response from cleverbot');
+        }
         bot.sendChat(response.message);
+      } else {
+        bot.log('error', 'BOT', 'No response from cleverbot');
       }
     });
   });